Refactored out the OpenIE features from ShallowNLPProcessor into a trait to be mixed into ShallowNLPProcessor's subclasses
