+++ corenlp/src/main/scala/org/clulab/processors/shallownlp/ShallowNLPProcessor.scala
@@ -46,7 +46,7 @@ class ShallowNLPProcessor(val tokenizerPostProcessor:Option[TokenizerStep],
   lazy val lemmatizer: StanfordCoreNLP = mkLemmatizer
   lazy val ner: StanfordCoreNLP = mkNer
   lazy val chunker: CRFChunker = mkChunker
-  lazy val openIE: StanfordCoreNLP = mkOpenIE
+
 
   protected def newStanfordCoreNLP(props: Properties, enforceRequirements: Boolean = true): StanfordCoreNLP = {
     // Prevent knownLCWords from changing on us.  To be safe, this is added every time
@@ -80,11 +80,7 @@ class ShallowNLPProcessor(val tokenizerPostProcessor:Option[TokenizerStep],
     CRFChunker.load(gzis)
   }
 
-  def mkOpenIE: StanfordCoreNLP ={
-    val props = new Properties()
-    props.put("annotators", "natlog,openie")
-    newStanfordCoreNLP(props, enforceRequirements = false)
-  }
+
 
   def in(s:String):String = {
     if (internStrings) Processor.internString(s)
@@ -260,6 +256,10 @@ class ShallowNLPProcessor(val tokenizerPostProcessor:Option[TokenizerStep],
     // nothing here; see classes that extend this
   }
 
+  override def relationExtraction(doc: Document): Unit  = {
+    // nothing here; see classes that extend this
+  }
+
   def chunking(doc:Document) {
     if (withChunks) {
       for (s <- doc.sentences) {
@@ -279,58 +279,6 @@ class ShallowNLPProcessor(val tokenizerPostProcessor:Option[TokenizerStep],
     // nothing here; see classes that extend this
   }
 
-  def relationExtractionSanityCheck(doc:Document):Option[Annotation] = {
-    if(withRelationExtraction) {
-      val annotation = basicSanityCheck(doc)
-      if (annotation.isEmpty) return None
-      if (doc.sentences.head.tags.isEmpty)
-        throw new RuntimeException("ERROR: you have to run the POS tagger before OpenIE!")
-      if (doc.sentences.head.lemmas.isEmpty)
-        throw new RuntimeException("ERROR: you have to run the lemmatizer before OpenIE!")
-      if (doc.sentences.head.dependencies.isEmpty)
-        throw new RuntimeException("ERROR: you have to run the dependency parser before OpenIE!")
-      annotation
-    }
-    else
-      None
-  }
-
-  def relationExtraction(doc: Document) {
-    val annotation = relationExtractionSanityCheck(doc)
-    if(annotation.isEmpty) return
-
-    try{
-      openIE.annotate(annotation.get)
-    } catch {
-      case e:Exception =>
-        println("Caught OpenIE exception!")
-        println("Document:\n" + doc)
-        throw e
-    }
-
-    // convert CoreNLP Annotations to our data structures
-    val sas = annotation.get.get(classOf[SentencesAnnotation]).asScala
-    var offset = 0
-    for (sa <- sas) {
-      val relationInstances = new ArrayBuffer[RelationTriple]
-      val triplets = sa.get(classOf[RelationTriplesAnnotation]).asScala
-      for (triplet <- triplets) {
-        val confidence = triplet.confidence
-        val subjectStart = triplet.canonicalSubject.asScala.head.index - 1
-        val subjectEnd = triplet.canonicalSubject.asScala.last.index
-        val relationStart = triplet.relation.asScala.head.index - 1
-        val relationEnd = triplet.relation.asScala.last.index
-        val objectStart = triplet.canonicalObject.asScala.head.index - 1
-        val objectEnd= triplet.canonicalObject.asScala.last.index
-
-        val relation = RelationTriple(confidence.toFloat, Interval(subjectStart, subjectEnd), Interval(relationStart, relationEnd), Interval(objectStart, objectEnd))
-        relationInstances += relation
-      }
-
-      doc.sentences(offset).relations = Some(relationInstances.toArray)
-      offset += 1
-    }
-  }
 }
 
 object ShallowNLPProcessor {
